// cloudapp/src/app/models/settings.ts
export class Settings {
  fontFamily: string = 'Arial';
  fontSize: string = '10pt';
  fontWeight: string = 'normal';
  lineHeight: string = '1';
  labelWidth: string = '.75in';
  labelHeight: string = '1in';
  marginLeft: string = '0.1in';
  marginRight: string = '0.1in';
  marginTop: string = '0.1in';
  marginBottom: string = '0.1in';
}